test: modularize test files
