+++ tests/unit/client.test.js
@@ -22,234 +22,10 @@ use(chaiAsPromised)
 
 import Nock from 'nock'
 
-import { CopyDestinationOptions, CopySourceOptions } from '../../src/helpers.ts'
-import {
-  calculateEvenSplits,
-  isValidEndpoint,
-  isValidIP,
-  makeDateLong,
-  makeDateShort,
-  partsRequired,
-} from '../../src/internal/helper.ts'
-import { joinHostPort } from '../../src/internal/join-host-port.ts'
+import { CopyDestinationOptions } from '../../src/helpers.ts'
 import * as Minio from '../../src/minio.js'
-import { parseListObjects } from '../../src/xml-parsers.js'
 
 const Package = { version: 'development' }
-
-describe('Helpers', () => {
-  it('should validate for s3 endpoint', () => {
-    assert.equal(isValidEndpoint('s3.amazonaws.com'), true)
-  })
-  it('should validate for s3 china', () => {
-    assert.equal(isValidEndpoint('s3.cn-north-1.amazonaws.com.cn'), true)
-  })
-  it('should validate for us-west-2', () => {
-    assert.equal(isValidEndpoint('s3-us-west-2.amazonaws.com'), true)
-  })
-  it('should fail for invalid endpoint characters', () => {
-    assert.equal(isValidEndpoint('111.#2.11'), false)
-  })
-
-  it('should make date short', () => {
-    let date = new Date('2012-12-03T17:25:36.331Z')
-
-    assert.equal(makeDateShort(date), '20121203')
-  })
-  it('should make date long', () => {
-    let date = new Date('2017-08-11T17:26:34.935Z')
-
-    assert.equal(makeDateLong(date), '20170811T172634Z')
-  })
-
-  // Adopted from minio-go sdk
-  const oneGB = 1024 * 1024 * 1024
-  const fiveGB = 5 * oneGB
-
-  const OBJ_SIZES = {
-    gb1: oneGB,
-    gb5: fiveGB,
-    gb5p1: fiveGB + 1,
-    gb10p1: 2 * fiveGB + 1,
-    gb10p2: 2 * fiveGB + 2,
-  }
-
-  const maxMultipartPutObjectSize = 1024 * 1024 * 1024 * 1024 * 5
-
-  it('Parts Required Test cases ', () => {
-    const expectedPartsRequiredTestCases = [
-      { value: 0, expected: 0 },
-      { value: 1, expected: 1 },
-      {
-        value: fiveGB,
-        expected: 10,
-      },
-      { value: OBJ_SIZES.gb5p1, expected: 10 },
-      { value: 2 * fiveGB, expected: 20 },
-      {
-        value: OBJ_SIZES.gb10p1,
-        expected: 20,
-      },
-      { value: OBJ_SIZES.gb10p2, expected: 20 },
-      {
-        value: OBJ_SIZES.gb10p1 + OBJ_SIZES.gb10p2,
-        expected: 40,
-      },
-      { value: maxMultipartPutObjectSize, expected: 10000 },
-    ]
-
-    expectedPartsRequiredTestCases.forEach((testCase) => {
-      const fnResult = partsRequired(testCase.value)
-      assert.equal(fnResult, testCase.expected)
-    })
-  })
-  it('Even split of Sizes Test cases ', () => {
-    // Adopted from minio-go sdk
-    const expectedSplitsTestCases = [
-      {
-        size: 0,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: null,
-        expectedEnd: null,
-      },
-      {
-        size: 1,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [undefined],
-        expectedEnd: [NaN],
-      },
-      { size: 1, sourceConfig: new CopySourceOptions({ Start: 0 }), expectedStart: [0], expectedEnd: [0] },
-      {
-        size: OBJ_SIZES.gb1,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [0, 536870912],
-        expectedEnd: [536870911, 1073741823],
-      },
-      {
-        size: OBJ_SIZES.gb5,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [
-          0, 536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
-        ],
-        expectedEnd: [
-          536870911, 1073741823, 1610612735, 2147483647, 2684354559, 3221225471, 3758096383, 4294967295, 4831838207,
-          5368709119,
-        ],
-      },
-
-      // 2 part splits
-      {
-        size: OBJ_SIZES.gb5p1,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [
-          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
-        ],
-        expectedEnd: [
-          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
-          5368709120,
-        ],
-      },
-      {
-        size: OBJ_SIZES.gb5p1,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [
-          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
-        ],
-        expectedEnd: [
-          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
-          5368709120,
-        ],
-      },
-
-      // 3 part splits
-      {
-        size: OBJ_SIZES.gb10p1,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [
-          0, 536870913, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
-          5368709121, 5905580033, 6442450945, 6979321857, 7516192769, 8053063681, 8589934593, 9126805505, 9663676417,
-          10200547329,
-        ],
-        expectedEnd: [
-          536870912, 1073741824, 1610612736, 2147483648, 2684354560, 3221225472, 3758096384, 4294967296, 4831838208,
-          5368709120, 5905580032, 6442450944, 6979321856, 7516192768, 8053063680, 8589934592, 9126805504, 9663676416,
-          10200547328, 10737418240,
-        ],
-      },
-      {
-        size: OBJ_SIZES.gb10p2,
-        sourceConfig: new CopySourceOptions({ Start: -1 }),
-        expectedStart: [
-          0, 536870913, 1073741826, 1610612738, 2147483650, 2684354562, 3221225474, 3758096386, 4294967298, 4831838210,
-          5368709122, 5905580034, 6442450946, 6979321858, 7516192770, 8053063682, 8589934594, 9126805506, 9663676418,
-          10200547330,
-        ],
-        expectedEnd: [
-          536870912, 1073741825, 1610612737, 2147483649, 2684354561, 3221225473, 3758096385, 4294967297, 4831838209,
-          5368709121, 5905580033, 6442450945, 6979321857, 7516192769, 8053063681, 8589934593, 9126805505, 9663676417,
-          10200547329, 10737418241,
-        ],
-      },
-    ]
-
-    expectedSplitsTestCases.forEach((testCase) => {
-      const fnResult = calculateEvenSplits(testCase.size, testCase)
-      const { startIndex, endIndex } = fnResult || {}
-
-      if (Array.isArray(startIndex) && Array.isArray(endIndex)) {
-        const isExpectedResult =
-          startIndex.length === testCase.expectedStart.length && endIndex.length === testCase.expectedEnd.length
-        assert.equal(isExpectedResult, true)
-      } else {
-        // null cases.
-        assert.equal(startIndex, expectedSplitsTestCases.expectedStart)
-        assert.equal(endIndex, expectedSplitsTestCases.expectedEnd)
-      }
-    })
-  })
-})
-
-describe('CopyConditions', () => {
-  let date = 'Fri, 11 Aug 2017 19:34:18 GMT'
-
-  let cc = new Minio.CopyConditions()
-
-  describe('#setModified', () => {
-    it('should take a date argument', () => {
-      cc.setModified(new Date(date))
-
-      assert.equal(cc.modified, date)
-    })
-
-    it('should throw without date', () => {
-      assert.throws(() => {
-        cc.setModified()
-      }, /date must be of type Date/)
-
-      assert.throws(() => {
-        cc.setModified({ hi: 'there' })
-      }, /date must be of type Date/)
-    })
-  })
-
-  describe('#setUnmodified', () => {
-    it('should take a date argument', () => {
-      cc.setUnmodified(new Date(date))
-
-      assert.equal(cc.unmodified, date)
-    })
-
-    it('should throw without date', () => {
-      assert.throws(() => {
-        cc.setUnmodified()
-      }, /date must be of type Date/)
-
-      assert.throws(() => {
-        cc.setUnmodified({ hi: 'there' })
-      }, /date must be of type Date/)
-    })
-  })
-})
 
 describe('Client', function () {
   var nockRequests = []
@@ -272,6 +48,7 @@ describe('Client', function () {
     secretKey: 'secretkey',
     useSSL: false,
   })
+
   describe('new client', () => {
     it('should work with https', () => {
       var client = new Minio.Client({
@@ -380,6 +157,7 @@ describe('Client', function () {
       }
     })
   })
+
   describe('Presigned URL', () => {
     describe('presigned-get', () => {
       it('should not generate presigned url with no access key', (done) => {
@@ -456,6 +234,7 @@ describe('Client', function () {
       })
     })
   })
+
   describe('User Agent', () => {
     it('should have a default user agent', () => {
       var client = new Minio.Client({
@@ -1053,6 +832,7 @@ describe('Client', function () {
       }
     })
   })
+
   describe('removeBucketLifecycle(bucket, callback)', () => {
     it('should fail on null bucket', (done) => {
       try {
@@ -1347,6 +1127,7 @@ describe('Client', function () {
       })
     })
   })
+
   describe('Bucket Replication APIs', () => {
     describe('setBucketReplication(bucketName, replicationConfig, callback)', () => {
       it('should fail on null bucket', (done) => {
@@ -1643,6 +1424,7 @@ describe('Client', function () {
       })
     })
   })
+
   describe('Select Object Content APIs', () => {
     describe('selectObjectContent(bucketName, objectName, selectOpts={}, cb)', () => {
       it('should fail on null bucket', (done) => {
@@ -1675,621 +1457,5 @@ describe('Client', function () {
         }
       })
     })
-  })
-})
-
-describe('IP Address Validations', () => {
-  it('should validate for valid ip', () => {
-    assert.equal(isValidIP('1.1.1.1'), true)
-  })
-
-  it('Check list of IPV4 Invalid addresses', () => {
-    const invalidIpv4 = [
-      ' 127.0.0.1',
-      '127.0.0.1 ',
-      '127.0.0.1 127.0.0.1',
-      '127.0.0.256',
-      '127.0.0.1//1',
-      '127.0.0.1/0x1',
-      '127.0.0.1/-1',
-      '127.0.0.1/ab',
-      '127.0.0.1/',
-      '127.0.0.256/32',
-      '127.0.0.1/33',
-    ]
-    invalidIpv4.map((ip) => {
-      assert.equal(isValidIP(ip), false)
-    })
-  })
-
-  it('Check list of IPV4 Valid addresses', () => {
-    const validIpv4 = ['001.002.003.004', '127.0.0.1', '255.255.255.255', '192.168.1.10']
-    validIpv4.map((ip) => {
-      assert.equal(isValidIP(ip), true)
-    })
-  })
-
-  it('Check list of IPV6 Invalid addresses', () => {
-    const invalidIpV6 = [
-      "':10.0.0.1",
-      '-1',
-      '::1 ::1',
-      '1.2.3.4:1111:2222:3333:4444::5555',
-      '1.2.3.4:1111:2222:3333::5555',
-      '1.2.3.4:1111:2222::5555',
-      '1.2.3.4:1111::5555',
-      '1.2.3.4::',
-      '1.2.3.4::5555',
-      '11112222:3333:4444:5555:6666:1.2.3.4',
-      '11112222:3333:4444:5555:6666:7777:8888',
-      '::1//64',
-      '::1/0001',
-      '1111:',
-      '1111:1.2.3.4',
-      '1111:2222',
-      '1111:22223333:4444:5555:6666:1.2.3.4',
-      '1111:22223333:4444:5555:6666:7777:8888',
-      '1111:2222:',
-      '1111:2222:1.2.3.4',
-      '1111:2222:3333',
-      '1111:2222:33334444:5555:6666:1.2.3.4',
-      '1111:2222:33334444:5555:6666:7777:8888',
-      '1111:2222:3333:',
-      '1111:2222:3333:1.2.3.4',
-      '1111:2222:3333:4444',
-      '1111:2222:3333:44445555:6666:1.2.3.4',
-      '1111:2222:3333:44445555:6666:7777:8888',
-      '1111:2222:3333:4444:',
-      '1111:2222:3333:4444:1.2.3.4',
-      '1111:2222:3333:4444:5555',
-      '1111:2222:3333:4444:55556666:1.2.3.4',
-      '1111:2222:3333:4444:55556666:7777:8888',
-      '1111:2222:3333:4444:5555:',
-      '1111:2222:3333:4444:5555:1.2.3.4',
-      '1111:2222:3333:4444:5555:6666',
-      '1111:2222:3333:4444:5555:66661.2.3.4',
-      '1111:2222:3333:4444:5555:66667777:8888',
-      '1111:2222:3333:4444:5555:6666:',
-      '1111:2222:3333:4444:5555:6666:1.2.3.4.5',
-      '1111:2222:3333:4444:5555:6666:255.255.255255',
-      '1111:2222:3333:4444:5555:6666:255.255255.255',
-      '1111:2222:3333:4444:5555:6666:255255.255.255',
-      '1111:2222:3333:4444:5555:6666:256.256.256.256',
-      '1111:2222:3333:4444:5555:6666:7777',
-      '1111:2222:3333:4444:5555:6666:77778888',
-      '1111:2222:3333:4444:5555:6666:7777:',
-      '1111:2222:3333:4444:5555:6666:7777:1.2.3.4',
-      '1111:2222:3333:4444:5555:6666:7777:::',
-      '1111:2222:3333:4444:5555:6666::8888:',
-      '1111:2222:3333:4444:5555:6666:::',
-      '1111:2222:3333:4444:5555:6666:::8888',
-      '1111:2222:3333:4444:5555::7777:8888:',
-      '1111:2222:3333:4444:5555::7777::',
-      '1111:2222:3333:4444:5555::8888:',
-      '1111:2222:3333:4444:5555:::',
-      '1111:2222:3333:4444:5555:::1.2.3.4',
-      '1111:2222:3333:4444:5555:::7777:8888',
-      '1111:2222:3333:4444::5555:',
-      '1111:2222:3333:4444::6666:7777:8888:',
-      '1111:2222:3333:4444::6666:7777::',
-      '1111:2222:3333:4444::6666::8888',
-      '1111:2222:3333:4444::7777:8888:',
-      '1111:2222:3333:4444::8888:',
-      '1111:2222:3333:4444:::',
-      '1111:2222:3333:4444:::6666:1.2.3.4',
-      '1111:2222:3333:4444:::6666:7777:8888',
-      '1111:2222:3333::5555:',
-      '1111:2222:3333::5555:6666:7777:8888:',
-      '1111:2222:3333::5555:6666:7777::',
-      '1111:2222:3333::5555:6666::8888',
-      '1111:2222:3333::5555::1.2.3.4',
-      '1111:2222:3333::5555::7777:8888',
-      '1111:2222:3333::6666:7777:8888:',
-      '1111:2222:3333::7777:8888:',
-      '1111:2222:3333::8888:',
-      '1111:2222:3333:::',
-      '1111:2222:3333:::5555:6666:1.2.3.4',
-      '1111:2222:3333:::5555:6666:7777:8888',
-      '1111:2222::4444:5555:6666:7777:8888:',
-      '1111:2222::4444:5555:6666:7777::',
-      '1111:2222::4444:5555:6666::8888',
-      '1111:2222::4444:5555::1.2.3.4',
-      '1111:2222::4444:5555::7777:8888',
-      '1111:2222::4444::6666:1.2.3.4',
-      '1111:2222::4444::6666:7777:8888',
-      '1111:2222::5555:',
-      '1111:2222::5555:6666:7777:8888:',
-      '1111:2222::6666:7777:8888:',
-      '1111:2222::7777:8888:',
-      '1111:2222::8888:',
-      '1111:2222:::',
-      '1111:2222:::4444:5555:6666:1.2.3.4',
-      '1111:2222:::4444:5555:6666:7777:8888',
-      '1111::3333:4444:5555:6666:7777:8888:',
-      '1111::3333:4444:5555:6666:7777::',
-      '1111::3333:4444:5555:6666::8888',
-      '1111::3333:4444:5555::1.2.3.4',
-      '1111::3333:4444:5555::7777:8888',
-      '1111::3333:4444::6666:1.2.3.4',
-      '1111::3333:4444::6666:7777:8888',
-      '1111::3333::5555:6666:1.2.3.4',
-      '1111::3333::5555:6666:7777:8888',
-      '1111::4444:5555:6666:7777:8888:',
-      '1111::5555:',
-      '1111::5555:6666:7777:8888:',
-      '1111::6666:7777:8888:',
-      '1111::7777:8888:',
-      '1111::8888:',
-      '1111:::',
-      '1111:::3333:4444:5555:6666:1.2.3.4',
-      '1111:::3333:4444:5555:6666:7777:8888',
-      '12345::6:7:8',
-      '124.15.6.89/60',
-      '1:2:3:4:5:6:7:8:9',
-      '1:2:3::4:5:6:7:8:9',
-      '1:2:3::4:5::7:8',
-      '1::1.2.256.4',
-      '1::1.2.3.256',
-      '1::1.2.3.300',
-      '1::1.2.3.900',
-      '1::1.2.300.4',
-      '1::1.2.900.4',
-      '1::1.256.3.4',
-      '1::1.300.3.4',
-      '1::1.900.3.4',
-      '1::256.2.3.4',
-      '1::260.2.3.4',
-      '1::2::3',
-      '1::300.2.3.4',
-      '1::300.300.300.300',
-      '1::3000.30.30.30',
-      '1::400.2.3.4',
-      '1::5:1.2.256.4',
-      '1::5:1.2.3.256',
-      '1::5:1.2.3.300',
-      '1::5:1.2.3.900',
-      '1::5:1.2.300.4',
-      '1::5:1.2.900.4',
-      '1::5:1.256.3.4',
-      '1::5:1.300.3.4',
-      '1::5:1.900.3.4',
-      '1::5:256.2.3.4',
-      '1::5:260.2.3.4',
-      '1::5:300.2.3.4',
-      '1::5:300.300.300.300',
-      '1::5:3000.30.30.30',
-      '1::5:400.2.3.4',
-      '1::5:900.2.3.4',
-      '1::900.2.3.4',
-      '1:::3:4:5',
-      '2001:0000:1234: 0000:0000:C1C0:ABCD:0876',
-      '2001:0000:1234:0000:0000:C1C0:ABCD:0876  0',
-      '2001:1:1:1:1:1:255Z255X255Y255',
-      '2001::FFD3::57ab',
-      '2001:DB8:0:0:8:800:200C:417A:221',
-      '2001:db8:85a3::8a2e:37023:7334',
-      '2001:db8:85a3::8a2e:370k:7334',
-      '3ffe:0b00:0000:0001:0000:0000:000a',
-      '3ffe:b00::1::a',
-      ':',
-      ':1.2.3.4',
-      ':1111:2222:3333:4444:5555:6666:1.2.3.4',
-      ':1111:2222:3333:4444:5555:6666:7777:8888',
-      ':1111:2222:3333:4444:5555:6666:7777::',
-      ':1111:2222:3333:4444:5555:6666::',
-      ':1111:2222:3333:4444:5555:6666::8888',
-      ':1111:2222:3333:4444:5555::',
-      ':1111:2222:3333:4444:5555::1.2.3.4',
-      ':1111:2222:3333:4444:5555::7777:8888',
-      ':1111:2222:3333:4444:5555::8888',
-      ':1111:2222:3333:4444::',
-      ':1111:2222:3333:4444::1.2.3.4',
-      ':1111:2222:3333:4444::5555',
-      ':1111:2222:3333:4444::6666:1.2.3.4',
-      ':1111:2222:3333:4444::6666:7777:8888',
-      ':1111:2222:3333:4444::7777:8888',
-      ':1111:2222:3333:4444::8888',
-      ':1111:2222:3333::',
-      ':1111:2222:3333::1.2.3.4',
-      ':1111:2222:3333::5555',
-      ':1111:2222:3333::5555:6666:1.2.3.4',
-      ':1111:2222:3333::5555:6666:7777:8888',
-      ':1111:2222:3333::6666:1.2.3.4',
-      ':1111:2222:3333::6666:7777:8888',
-      ':1111:2222:3333::7777:8888',
-      ':1111:2222:3333::8888',
-      ':1111:2222::',
-      ':1111:2222::1.2.3.4',
-      ':1111:2222::4444:5555:6666:1.2.3.4',
-      ':1111:2222::4444:5555:6666:7777:8888',
-      ':1111:2222::5555',
-      ':1111:2222::5555:6666:1.2.3.4',
-      ':1111:2222::5555:6666:7777:8888',
-      ':1111:2222::6666:1.2.3.4',
-      ':1111:2222::6666:7777:8888',
-      ':1111:2222::7777:8888',
-      ':1111:2222::8888',
-      ':1111::',
-      ':1111::1.2.3.4',
-      ':1111::3333:4444:5555:6666:1.2.3.4',
-      ':1111::3333:4444:5555:6666:7777:8888',
-      ':1111::4444:5555:6666:1.2.3.4',
-      ':1111::4444:5555:6666:7777:8888',
-      ':1111::5555',
-      ':1111::5555:6666:1.2.3.4',
-      ':1111::5555:6666:7777:8888',
-      ':1111::6666:1.2.3.4',
-      ':1111::6666:7777:8888',
-      ':1111::7777:8888',
-      ':1111::8888',
-      ':2222:3333:4444:5555:6666:1.2.3.4',
-      ':2222:3333:4444:5555:6666:7777:8888',
-      ':3333:4444:5555:6666:1.2.3.4',
-      ':3333:4444:5555:6666:7777:8888',
-      ':4444:5555:6666:1.2.3.4',
-      ':4444:5555:6666:7777:8888',
-      ':5555:6666:1.2.3.4',
-      ':5555:6666:7777:8888',
-      ':6666:1.2.3.4',
-      ':6666:7777:8888',
-      ':7777:8888',
-      ':8888',
-      '::-1',
-      '::.',
-      '::..',
-      '::...',
-      '::...4',
-      '::..3.',
-      '::..3.4',
-      '::.2..',
-      '::.2.3.',
-      '::.2.3.4',
-      '::1...',
-      '::1.2..',
-      '::1.2.256.4',
-      '::1.2.3.',
-      '::1.2.3.256',
-      '::1.2.3.300',
-      '::1.2.3.900',
-      '::1.2.300.4',
-      '::1.2.900.4',
-      '::1.256.3.4',
-      '::1.300.3.4',
-      '::1.900.3.4',
-      '::1111:2222:3333:4444:5555:6666::',
-      '::2222:3333:4444:5555:6666:7777:8888:',
-      '::2222:3333:4444:5555:7777:8888::',
-      '::2222:3333:4444:5555:7777::8888',
-      '::2222:3333:4444:5555::1.2.3.4',
-      '::2222:3333:4444:5555::7777:8888',
-      '::2222:3333:4444::6666:1.2.3.4',
-      '::2222:3333:4444::6666:7777:8888',
-      '::2222:3333::5555:6666:1.2.3.4',
-      '::2222:3333::5555:6666:7777:8888',
-      '::2222::4444:5555:6666:1.2.3.4',
-      '::2222::4444:5555:6666:7777:8888',
-      '::256.2.3.4',
-      '::260.2.3.4',
-      '::300.2.3.4',
-      '::300.300.300.300',
-      '::3000.30.30.30',
-      '::3333:4444:5555:6666:7777:8888:',
-      '::400.2.3.4',
-      '::4444:5555:6666:7777:8888:',
-      '::5555:',
-      '::5555:6666:7777:8888:',
-      '::6666:7777:8888:',
-      '::7777:8888:',
-      '::8888:',
-      '::900.2.3.4',
-      ':::',
-      ':::1.2.3.4',
-      ':::2222:3333:4444:5555:6666:1.2.3.4',
-      ':::2222:3333:4444:5555:6666:7777:8888',
-      ':::3333:4444:5555:6666:7777:8888',
-      ':::4444:5555:6666:1.2.3.4',
-      ':::4444:5555:6666:7777:8888',
-      ':::5555',
-      ':::5555:6666:1.2.3.4',
-      ':::5555:6666:7777:8888',
-      ':::6666:1.2.3.4',
-      ':::6666:7777:8888',
-      ':::7777:8888',
-      ':::8888',
-      '::ffff:192x168.1.26',
-      '::ffff:2.3.4',
-      '::ffff:257.1.2.3',
-      'FF01::101::2',
-      'FF02:0000:0000:0000:0000:0000:0000:0000:0001',
-      'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:1.2.3.4',
-      'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX',
-      'a::b::c',
-      'a::g',
-      'a:a:a:a:a:a:a:a:a',
-      'a:aaaaa::',
-      'a:b',
-      'a:b:c:d:e:f:g:0',
-      'ffff:',
-      'ffff::ffff::ffff',
-      'ffgg:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
-      'ldkfj',
-      '::/129',
-      '1000:://32',
-      '::/',
-    ]
-    invalidIpV6.map((ip) => {
-      const valid = isValidIP(ip)
-      assert.equal(valid, false)
-    })
-  })
-
-  it('Check list of IPV6 Valid addresses', () => {
-    const validIpv6 = [
-      '0000:0000:0000:0000:0000:0000:0000:0000',
-      '0000:0000:0000:0000:0000:0000:0000:0001',
-      '0:0:0:0:0:0:0:0',
-      '0:0:0:0:0:0:0:1',
-      '0:0:0:0:0:0:0::',
-      '0:0:0:0:0:0:13.1.68.3',
-      '0:0:0:0:0:0::',
-      '0:0:0:0:0::',
-      '0:0:0:0:0:FFFF:129.144.52.38',
-      '0:0:0:0:1:0:0:0',
-      '0:0:0:0::',
-      '0:0:0::',
-      '0:0::',
-      '0:1:2:3:4:5:6:7',
-      '0::',
-      '0:a:b:c:d:e:f::',
-      '1080:0:0:0:8:800:200c:417a',
-      '1080::8:800:200c:417a',
-      '1111:2222:3333:4444:5555:6666:123.123.123.123',
-      '1111:2222:3333:4444:5555:6666:7777:8888',
-      '1111:2222:3333:4444:5555:6666:7777::',
-      '1111:2222:3333:4444:5555:6666::',
-      '1111:2222:3333:4444:5555:6666::8888',
-      '1111:2222:3333:4444:5555::',
-      '1111:2222:3333:4444:5555::7777:8888',
-      '1111:2222:3333:4444:5555::8888',
-      '1111:2222:3333:4444::',
-      '1111:2222:3333:4444::6666:123.123.123.123',
-      '1111:2222:3333:4444::6666:7777:8888',
-      '1111:2222:3333:4444::7777:8888',
-      '1111:2222:3333:4444::8888',
-      '1111:2222:3333::',
-      '1111:2222:3333::5555:6666:123.123.123.123',
-      '1111:2222:3333::5555:6666:7777:8888',
-      '1111:2222:3333::6666:123.123.123.123',
-      '1111:2222:3333::6666:7777:8888',
-      '1111:2222:3333::7777:8888',
-      '1111:2222:3333::8888',
-      '1111:2222::',
-      '1111:2222::4444:5555:6666:123.123.123.123',
-      '1111:2222::4444:5555:6666:7777:8888',
-      '1111:2222::5555:6666:123.123.123.123',
-      '1111:2222::5555:6666:7777:8888',
-      '1111:2222::6666:123.123.123.123',
-      '1111:2222::6666:7777:8888',
-      '1111:2222::7777:8888',
-      '1111:2222::8888',
-      '1111::',
-      '1111::3333:4444:5555:6666:123.123.123.123',
-      '1111::3333:4444:5555:6666:7777:8888',
-      '1111::4444:5555:6666:123.123.123.123',
-      '1111::4444:5555:6666:7777:8888',
-      '1111::5555:6666:123.123.123.123',
-      '1111::5555:6666:7777:8888',
-      '1111::6666:123.123.123.123',
-      '1111::6666:7777:8888',
-      '1111::7777:8888',
-      '1111::8888',
-      '1:2:3:4:5:6:1.2.3.4',
-      '1:2:3:4:5:6:7:8',
-      '1:2:3:4:5:6::',
-      '1:2:3:4:5:6::8',
-      '1:2:3:4:5::',
-      '1:2:3:4:5::7:8',
-      '1:2:3:4:5::8',
-      '1:2:3:4::',
-      '1:2:3:4::5:1.2.3.4',
-      '1:2:3:4::7:8',
-      '1:2:3:4::8',
-      '1:2:3::',
-      '1:2:3::5:1.2.3.4',
-      '1:2:3::7:8',
-      '1:2:3::8',
-      '1:2::',
-      '1:2::5:1.2.3.4',
-      '1:2::7:8',
-      '1:2::8',
-      '1::',
-      '1::2:3',
-      '1::2:3:4',
-      '1::2:3:4:5',
-      '1::2:3:4:5:6',
-      '1::2:3:4:5:6:7',
-      '1::5:1.2.3.4',
-      '1::5:11.22.33.44',
-      '1::7:8',
-      '1::8',
-      '2001:0000:1234:0000:0000:C1C0:ABCD:0876',
-      '2001:0000:4136:e378:8000:63bf:3fff:fdd2',
-      '2001:0db8:0000:0000:0000:0000:1428:57ab',
-      '2001:0db8:0000:0000:0000::1428:57ab',
-      '2001:0db8:0:0:0:0:1428:57ab',
-      '2001:0db8:0:0::1428:57ab',
-      '2001:0db8:1234:0000:0000:0000:0000:0000',
-      '2001:0db8:1234::',
-      '2001:0db8:1234:ffff:ffff:ffff:ffff:ffff',
-      '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
-      '2001:0db8::1428:57ab',
-      '2001::CE49:7601:2CAD:DFFF:7C94:FFFE',
-      '2001::CE49:7601:E866:EFFF:62C3:FFFE',
-      '2001:DB8:0:0:8:800:200C:417A',
-      '2001:DB8::8:800:200C:417A',
-      '2001:db8:85a3:0:0:8a2e:370:7334',
-      '2001:db8:85a3::8a2e:370:7334',
-      '2001:db8::',
-      '2001:db8::1428:57ab',
-      '2001:db8:a::123',
-      '2002::',
-      '2608::3:5',
-      '2608:af09:30:0:0:0:0:134',
-      '2608:af09:30::102a:7b91:c239:baff',
-      '2::10',
-      '3ffe:0b00:0000:0000:0001:0000:0000:000a',
-      '7:6:5:4:3:2:1:0',
-      '::',
-      '::0',
-      '::0:0',
-      '::0:0:0',
-      '::0:0:0:0',
-      '::0:0:0:0:0',
-      '::0:0:0:0:0:0',
-      '::0:0:0:0:0:0:0',
-      '::0:a:b:c:d:e:f',
-      '::1',
-      '::123.123.123.123',
-      '::13.1.68.3',
-      '::2222:3333:4444:5555:6666:123.123.123.123',
-      '::2222:3333:4444:5555:6666:7777:8888',
-      '::2:3',
-      '::2:3:4',
-      '::2:3:4:5',
-      '::2:3:4:5:6',
-      '::2:3:4:5:6:7',
-      '::2:3:4:5:6:7:8',
-      '::3333:4444:5555:6666:7777:8888',
-      '::4444:5555:6666:123.123.123.123',
-      '::4444:5555:6666:7777:8888',
-      '::5555:6666:123.123.123.123',
-      '::5555:6666:7777:8888',
-      '::6666:123.123.123.123',
-      '::6666:7777:8888',
-      '::7777:8888',
-      '::8',
-      '::8888',
-      '::FFFF:129.144.52.38',
-      '::ffff:0:0',
-      '::ffff:0c22:384e',
-      '::ffff:12.34.56.78',
-      '::ffff:192.0.2.128',
-      '::ffff:192.168.1.1',
-      '::ffff:192.168.1.26',
-      '::ffff:c000:280',
-      'FF01:0:0:0:0:0:0:101',
-      'FF01::101',
-      'FF02:0000:0000:0000:0000:0000:0000:0001',
-      'a:b:c:d:e:f:0::',
-      'fe80:0000:0000:0000:0204:61ff:fe9d:f156',
-      'fe80:0:0:0:204:61ff:254.157.241.86',
-      'fe80:0:0:0:204:61ff:fe9d:f156',
-      'fe80::',
-      'fe80::1',
-      'fe80::204:61ff:254.157.241.86',
-      'fe80::204:61ff:fe9d:f156',
-      'fe80::217:f2ff:254.7.237.98',
-      'fe80::217:f2ff:fe07:ed62',
-      'fedc:ba98:7654:3210:fedc:ba98:7654:3210',
-      'ff02::1',
-      'ffff::',
-      'ffff::3:5',
-      'ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
-      'a:0::0:b',
-      'a:0:0::0:b',
-      'a:0::0:0:b',
-      'a::0:0:b',
-      'a::0:b',
-      'a:0::b',
-      'a:0:0::b',
-    ]
-    validIpv6.map((ip) => {
-      const valid = isValidIP(ip)
-      assert.equal(valid, true)
-    })
-  })
-})
-
-describe('xml-parser', () => {
-  describe('#listObjects()', () => {
-    describe('value type casting', () => {
-      const xml = `
-          <?xml version="1.0" encoding="UTF-8"?>
-          <ListVersionsResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-            <Name>some-bucket</Name>
-            <Prefix>42</Prefix>
-            <Delimiter>/</Delimiter>
-            <IsTruncated>false</IsTruncated>
-            <EncodingType>url</EncodingType>
-            <KeyMarker/>
-            <VersionIdMarker/>
-            <Version>
-              <IsLatest>true</IsLatest>
-              <VersionId>1234</VersionId>
-              <ETag>"767dedcb515a0e2d995ed95191b75484-29"</ETag>
-              <Key>1337</Key>
-              <LastModified>2023-07-12T14:41:46.000Z</LastModified>
-              <Size>151306240</Size>
-            </Version>
-            <DeleteMarker>
-              <IsLatest>false</IsLatest>
-              <Key>1337</Key>
-              <LastModified>2023-07-12T14:39:22.000Z</LastModified>
-              <VersionId>5678</VersionId>
-            </DeleteMarker>
-            <CommonPrefixes>
-              <Prefix>42</Prefix>
-            </CommonPrefixes>
-          </ListVersionsResult>
-        `
-
-      it('should parse VersionId as string even if number is provided', () => {
-        const { objects } = parseListObjects(xml)
-
-        assert.equal(objects[0].versionId, '1234')
-        assert.equal(objects[1].versionId, '5678')
-        assert.equal(objects[0].name, '1337')
-        assert.equal(objects[1].name, '1337')
-        assert.deepEqual(objects[2], { prefix: '42', size: 0 })
-      })
-
-      it('should parse Size as number', () => {
-        const { objects } = parseListObjects(xml)
-
-        assert.equal(objects[0].size, 151306240)
-        assert.equal(objects[1].size, undefined)
-      })
-    })
-  })
-})
-
-describe('join-host-port', () => {
-  it('should be able to parse valid ipv4', () => {
-    assert.equal(joinHostPort('192.168.1.1', 3000), '192.168.1.1:3000')
-    assert.equal(joinHostPort('192.168.1.1'), '192.168.1.1')
-    assert.equal(joinHostPort('01.102.103.104', 3000), '01.102.103.104:3000')
-  })
-
-  it('should be able to parse valid ipv6', () => {
-    assert.equal(
-      joinHostPort('2001:db8:3333:4444:5555:6666:7777:8888', 1234),
-      '[2001:db8:3333:4444:5555:6666:7777:8888]:1234',
-    )
-    assert.equal(joinHostPort('::', 1234), '[::]:1234')
-    assert.equal(joinHostPort('2001:db8::', 1234), '[2001:db8::]:1234')
-    assert.equal(joinHostPort('::1234:5678', 1234), '[::1234:5678]:1234')
-    assert.equal(
-      joinHostPort('2001:0db8:0001:0000:0000:0ab9:C0A8:0102', 1234),
-      '[2001:0db8:0001:0000:0000:0ab9:C0A8:0102]:1234',
-    )
-    assert.equal(
-      joinHostPort('2001:db8:3333:4444:5555:6666:1.2.3.4', 1234),
-      '[2001:db8:3333:4444:5555:6666:1.2.3.4]:1234',
-    )
-  })
-
-  it('should be able to parse domain', () => {
-    assert.equal(joinHostPort('internal.company.com', 5000), 'internal.company.com:5000')
-    assert.equal(joinHostPort('internal.company.com'), 'internal.company.com')
   })
 })
