+++ tests/unit/ip-address.test.js
@@ -1,0 +1,556 @@
+/*
+ * MinIO Javascript Library for Amazon S3 Compatible Cloud Storage, (C) 2016 MinIO, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { assert, use } from 'chai'
+import chaiAsPromised from 'chai-as-promised'
+use(chaiAsPromised)
+
+import {
+  isValidIP,
+} from '../../src/internal/helper.ts'
+
+describe('IP Address Validations', () => {
+    it('should validate for valid ip', () => {
+      assert.equal(isValidIP('1.1.1.1'), true)
+    })
+  
+    it('Check list of IPV4 Invalid addresses', () => {
+      const invalidIpv4 = [
+        ' 127.0.0.1',
+        '127.0.0.1 ',
+        '127.0.0.1 127.0.0.1',
+        '127.0.0.256',
+        '127.0.0.1//1',
+        '127.0.0.1/0x1',
+        '127.0.0.1/-1',
+        '127.0.0.1/ab',
+        '127.0.0.1/',
+        '127.0.0.256/32',
+        '127.0.0.1/33',
+      ]
+      invalidIpv4.map((ip) => {
+        assert.equal(isValidIP(ip), false)
+      })
+    })
+  
+    it('Check list of IPV4 Valid addresses', () => {
+      const validIpv4 = ['001.002.003.004', '127.0.0.1', '255.255.255.255', '192.168.1.10']
+      validIpv4.map((ip) => {
+        assert.equal(isValidIP(ip), true)
+      })
+    })
+  
+    it('Check list of IPV6 Invalid addresses', () => {
+      const invalidIpV6 = [
+        "':10.0.0.1",
+        '-1',
+        '::1 ::1',
+        '1.2.3.4:1111:2222:3333:4444::5555',
+        '1.2.3.4:1111:2222:3333::5555',
+        '1.2.3.4:1111:2222::5555',
+        '1.2.3.4:1111::5555',
+        '1.2.3.4::',
+        '1.2.3.4::5555',
+        '11112222:3333:4444:5555:6666:1.2.3.4',
+        '11112222:3333:4444:5555:6666:7777:8888',
+        '::1//64',
+        '::1/0001',
+        '1111:',
+        '1111:1.2.3.4',
+        '1111:2222',
+        '1111:22223333:4444:5555:6666:1.2.3.4',
+        '1111:22223333:4444:5555:6666:7777:8888',
+        '1111:2222:',
+        '1111:2222:1.2.3.4',
+        '1111:2222:3333',
+        '1111:2222:33334444:5555:6666:1.2.3.4',
+        '1111:2222:33334444:5555:6666:7777:8888',
+        '1111:2222:3333:',
+        '1111:2222:3333:1.2.3.4',
+        '1111:2222:3333:4444',
+        '1111:2222:3333:44445555:6666:1.2.3.4',
+        '1111:2222:3333:44445555:6666:7777:8888',
+        '1111:2222:3333:4444:',
+        '1111:2222:3333:4444:1.2.3.4',
+        '1111:2222:3333:4444:5555',
+        '1111:2222:3333:4444:55556666:1.2.3.4',
+        '1111:2222:3333:4444:55556666:7777:8888',
+        '1111:2222:3333:4444:5555:',
+        '1111:2222:3333:4444:5555:1.2.3.4',
+        '1111:2222:3333:4444:5555:6666',
+        '1111:2222:3333:4444:5555:66661.2.3.4',
+        '1111:2222:3333:4444:5555:66667777:8888',
+        '1111:2222:3333:4444:5555:6666:',
+        '1111:2222:3333:4444:5555:6666:1.2.3.4.5',
+        '1111:2222:3333:4444:5555:6666:255.255.255255',
+        '1111:2222:3333:4444:5555:6666:255.255255.255',
+        '1111:2222:3333:4444:5555:6666:255255.255.255',
+        '1111:2222:3333:4444:5555:6666:256.256.256.256',
+        '1111:2222:3333:4444:5555:6666:7777',
+        '1111:2222:3333:4444:5555:6666:77778888',
+        '1111:2222:3333:4444:5555:6666:7777:',
+        '1111:2222:3333:4444:5555:6666:7777:1.2.3.4',
+        '1111:2222:3333:4444:5555:6666:7777:::',
+        '1111:2222:3333:4444:5555:6666::8888:',
+        '1111:2222:3333:4444:5555:6666:::',
+        '1111:2222:3333:4444:5555:6666:::8888',
+        '1111:2222:3333:4444:5555::7777:8888:',
+        '1111:2222:3333:4444:5555::7777::',
+        '1111:2222:3333:4444:5555::8888:',
+        '1111:2222:3333:4444:5555:::',
+        '1111:2222:3333:4444:5555:::1.2.3.4',
+        '1111:2222:3333:4444:5555:::7777:8888',
+        '1111:2222:3333:4444::5555:',
+        '1111:2222:3333:4444::6666:7777:8888:',
+        '1111:2222:3333:4444::6666:7777::',
+        '1111:2222:3333:4444::6666::8888',
+        '1111:2222:3333:4444::7777:8888:',
+        '1111:2222:3333:4444::8888:',
+        '1111:2222:3333:4444:::',
+        '1111:2222:3333:4444:::6666:1.2.3.4',
+        '1111:2222:3333:4444:::6666:7777:8888',
+        '1111:2222:3333::5555:',
+        '1111:2222:3333::5555:6666:7777:8888:',
+        '1111:2222:3333::5555:6666:7777::',
+        '1111:2222:3333::5555:6666::8888',
+        '1111:2222:3333::5555::1.2.3.4',
+        '1111:2222:3333::5555::7777:8888',
+        '1111:2222:3333::6666:7777:8888:',
+        '1111:2222:3333::7777:8888:',
+        '1111:2222:3333::8888:',
+        '1111:2222:3333:::',
+        '1111:2222:3333:::5555:6666:1.2.3.4',
+        '1111:2222:3333:::5555:6666:7777:8888',
+        '1111:2222::4444:5555:6666:7777:8888:',
+        '1111:2222::4444:5555:6666:7777::',
+        '1111:2222::4444:5555:6666::8888',
+        '1111:2222::4444:5555::1.2.3.4',
+        '1111:2222::4444:5555::7777:8888',
+        '1111:2222::4444::6666:1.2.3.4',
+        '1111:2222::4444::6666:7777:8888',
+        '1111:2222::5555:',
+        '1111:2222::5555:6666:7777:8888:',
+        '1111:2222::6666:7777:8888:',
+        '1111:2222::7777:8888:',
+        '1111:2222::8888:',
+        '1111:2222:::',
+        '1111:2222:::4444:5555:6666:1.2.3.4',
+        '1111:2222:::4444:5555:6666:7777:8888',
+        '1111::3333:4444:5555:6666:7777:8888:',
+        '1111::3333:4444:5555:6666:7777::',
+        '1111::3333:4444:5555:6666::8888',
+        '1111::3333:4444:5555::1.2.3.4',
+        '1111::3333:4444:5555::7777:8888',
+        '1111::3333:4444::6666:1.2.3.4',
+        '1111::3333:4444::6666:7777:8888',
+        '1111::3333::5555:6666:1.2.3.4',
+        '1111::3333::5555:6666:7777:8888',
+        '1111::4444:5555:6666:7777:8888:',
+        '1111::5555:',
+        '1111::5555:6666:7777:8888:',
+        '1111::6666:7777:8888:',
+        '1111::7777:8888:',
+        '1111::8888:',
+        '1111:::',
+        '1111:::3333:4444:5555:6666:1.2.3.4',
+        '1111:::3333:4444:5555:6666:7777:8888',
+        '12345::6:7:8',
+        '124.15.6.89/60',
+        '1:2:3:4:5:6:7:8:9',
+        '1:2:3::4:5:6:7:8:9',
+        '1:2:3::4:5::7:8',
+        '1::1.2.256.4',
+        '1::1.2.3.256',
+        '1::1.2.3.300',
+        '1::1.2.3.900',
+        '1::1.2.300.4',
+        '1::1.2.900.4',
+        '1::1.256.3.4',
+        '1::1.300.3.4',
+        '1::1.900.3.4',
+        '1::256.2.3.4',
+        '1::260.2.3.4',
+        '1::2::3',
+        '1::300.2.3.4',
+        '1::300.300.300.300',
+        '1::3000.30.30.30',
+        '1::400.2.3.4',
+        '1::5:1.2.256.4',
+        '1::5:1.2.3.256',
+        '1::5:1.2.3.300',
+        '1::5:1.2.3.900',
+        '1::5:1.2.300.4',
+        '1::5:1.2.900.4',
+        '1::5:1.256.3.4',
+        '1::5:1.300.3.4',
+        '1::5:1.900.3.4',
+        '1::5:256.2.3.4',
+        '1::5:260.2.3.4',
+        '1::5:300.2.3.4',
+        '1::5:300.300.300.300',
+        '1::5:3000.30.30.30',
+        '1::5:400.2.3.4',
+        '1::5:900.2.3.4',
+        '1::900.2.3.4',
+        '1:::3:4:5',
+        '2001:0000:1234: 0000:0000:C1C0:ABCD:0876',
+        '2001:0000:1234:0000:0000:C1C0:ABCD:0876  0',
+        '2001:1:1:1:1:1:255Z255X255Y255',
+        '2001::FFD3::57ab',
+        '2001:DB8:0:0:8:800:200C:417A:221',
+        '2001:db8:85a3::8a2e:37023:7334',
+        '2001:db8:85a3::8a2e:370k:7334',
+        '3ffe:0b00:0000:0001:0000:0000:000a',
+        '3ffe:b00::1::a',
+        ':',
+        ':1.2.3.4',
+        ':1111:2222:3333:4444:5555:6666:1.2.3.4',
+        ':1111:2222:3333:4444:5555:6666:7777:8888',
+        ':1111:2222:3333:4444:5555:6666:7777::',
+        ':1111:2222:3333:4444:5555:6666::',
+        ':1111:2222:3333:4444:5555:6666::8888',
+        ':1111:2222:3333:4444:5555::',
+        ':1111:2222:3333:4444:5555::1.2.3.4',
+        ':1111:2222:3333:4444:5555::7777:8888',
+        ':1111:2222:3333:4444:5555::8888',
+        ':1111:2222:3333:4444::',
+        ':1111:2222:3333:4444::1.2.3.4',
+        ':1111:2222:3333:4444::5555',
+        ':1111:2222:3333:4444::6666:1.2.3.4',
+        ':1111:2222:3333:4444::6666:7777:8888',
+        ':1111:2222:3333:4444::7777:8888',
+        ':1111:2222:3333:4444::8888',
+        ':1111:2222:3333::',
+        ':1111:2222:3333::1.2.3.4',
+        ':1111:2222:3333::5555',
+        ':1111:2222:3333::5555:6666:1.2.3.4',
+        ':1111:2222:3333::5555:6666:7777:8888',
+        ':1111:2222:3333::6666:1.2.3.4',
+        ':1111:2222:3333::6666:7777:8888',
+        ':1111:2222:3333::7777:8888',
+        ':1111:2222:3333::8888',
+        ':1111:2222::',
+        ':1111:2222::1.2.3.4',
+        ':1111:2222::4444:5555:6666:1.2.3.4',
+        ':1111:2222::4444:5555:6666:7777:8888',
+        ':1111:2222::5555',
+        ':1111:2222::5555:6666:1.2.3.4',
+        ':1111:2222::5555:6666:7777:8888',
+        ':1111:2222::6666:1.2.3.4',
+        ':1111:2222::6666:7777:8888',
+        ':1111:2222::7777:8888',
+        ':1111:2222::8888',
+        ':1111::',
+        ':1111::1.2.3.4',
+        ':1111::3333:4444:5555:6666:1.2.3.4',
+        ':1111::3333:4444:5555:6666:7777:8888',
+        ':1111::4444:5555:6666:1.2.3.4',
+        ':1111::4444:5555:6666:7777:8888',
+        ':1111::5555',
+        ':1111::5555:6666:1.2.3.4',
+        ':1111::5555:6666:7777:8888',
+        ':1111::6666:1.2.3.4',
+        ':1111::6666:7777:8888',
+        ':1111::7777:8888',
+        ':1111::8888',
+        ':2222:3333:4444:5555:6666:1.2.3.4',
+        ':2222:3333:4444:5555:6666:7777:8888',
+        ':3333:4444:5555:6666:1.2.3.4',
+        ':3333:4444:5555:6666:7777:8888',
+        ':4444:5555:6666:1.2.3.4',
+        ':4444:5555:6666:7777:8888',
+        ':5555:6666:1.2.3.4',
+        ':5555:6666:7777:8888',
+        ':6666:1.2.3.4',
+        ':6666:7777:8888',
+        ':7777:8888',
+        ':8888',
+        '::-1',
+        '::.',
+        '::..',
+        '::...',
+        '::...4',
+        '::..3.',
+        '::..3.4',
+        '::.2..',
+        '::.2.3.',
+        '::.2.3.4',
+        '::1...',
+        '::1.2..',
+        '::1.2.256.4',
+        '::1.2.3.',
+        '::1.2.3.256',
+        '::1.2.3.300',
+        '::1.2.3.900',
+        '::1.2.300.4',
+        '::1.2.900.4',
+        '::1.256.3.4',
+        '::1.300.3.4',
+        '::1.900.3.4',
+        '::1111:2222:3333:4444:5555:6666::',
+        '::2222:3333:4444:5555:6666:7777:8888:',
+        '::2222:3333:4444:5555:7777:8888::',
+        '::2222:3333:4444:5555:7777::8888',
+        '::2222:3333:4444:5555::1.2.3.4',
+        '::2222:3333:4444:5555::7777:8888',
+        '::2222:3333:4444::6666:1.2.3.4',
+        '::2222:3333:4444::6666:7777:8888',
+        '::2222:3333::5555:6666:1.2.3.4',
+        '::2222:3333::5555:6666:7777:8888',
+        '::2222::4444:5555:6666:1.2.3.4',
+        '::2222::4444:5555:6666:7777:8888',
+        '::256.2.3.4',
+        '::260.2.3.4',
+        '::300.2.3.4',
+        '::300.300.300.300',
+        '::3000.30.30.30',
+        '::3333:4444:5555:6666:7777:8888:',
+        '::400.2.3.4',
+        '::4444:5555:6666:7777:8888:',
+        '::5555:',
+        '::5555:6666:7777:8888:',
+        '::6666:7777:8888:',
+        '::7777:8888:',
+        '::8888:',
+        '::900.2.3.4',
+        ':::',
+        ':::1.2.3.4',
+        ':::2222:3333:4444:5555:6666:1.2.3.4',
+        ':::2222:3333:4444:5555:6666:7777:8888',
+        ':::3333:4444:5555:6666:7777:8888',
+        ':::4444:5555:6666:1.2.3.4',
+        ':::4444:5555:6666:7777:8888',
+        ':::5555',
+        ':::5555:6666:1.2.3.4',
+        ':::5555:6666:7777:8888',
+        ':::6666:1.2.3.4',
+        ':::6666:7777:8888',
+        ':::7777:8888',
+        ':::8888',
+        '::ffff:192x168.1.26',
+        '::ffff:2.3.4',
+        '::ffff:257.1.2.3',
+        'FF01::101::2',
+        'FF02:0000:0000:0000:0000:0000:0000:0000:0001',
+        'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:1.2.3.4',
+        'XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX:XXXX',
+        'a::b::c',
+        'a::g',
+        'a:a:a:a:a:a:a:a:a',
+        'a:aaaaa::',
+        'a:b',
+        'a:b:c:d:e:f:g:0',
+        'ffff:',
+        'ffff::ffff::ffff',
+        'ffgg:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
+        'ldkfj',
+        '::/129',
+        '1000:://32',
+        '::/',
+      ]
+      invalidIpV6.map((ip) => {
+        const valid = isValidIP(ip)
+        assert.equal(valid, false)
+      })
+    })
+  
+    it('Check list of IPV6 Valid addresses', () => {
+      const validIpv6 = [
+        '0000:0000:0000:0000:0000:0000:0000:0000',
+        '0000:0000:0000:0000:0000:0000:0000:0001',
+        '0:0:0:0:0:0:0:0',
+        '0:0:0:0:0:0:0:1',
+        '0:0:0:0:0:0:0::',
+        '0:0:0:0:0:0:13.1.68.3',
+        '0:0:0:0:0:0::',
+        '0:0:0:0:0::',
+        '0:0:0:0:0:FFFF:129.144.52.38',
+        '0:0:0:0:1:0:0:0',
+        '0:0:0:0::',
+        '0:0:0::',
+        '0:0::',
+        '0:1:2:3:4:5:6:7',
+        '0::',
+        '0:a:b:c:d:e:f::',
+        '1080:0:0:0:8:800:200c:417a',
+        '1080::8:800:200c:417a',
+        '1111:2222:3333:4444:5555:6666:123.123.123.123',
+        '1111:2222:3333:4444:5555:6666:7777:8888',
+        '1111:2222:3333:4444:5555:6666:7777::',
+        '1111:2222:3333:4444:5555:6666::',
+        '1111:2222:3333:4444:5555:6666::8888',
+        '1111:2222:3333:4444:5555::',
+        '1111:2222:3333:4444:5555::7777:8888',
+        '1111:2222:3333:4444:5555::8888',
+        '1111:2222:3333:4444::',
+        '1111:2222:3333:4444::6666:123.123.123.123',
+        '1111:2222:3333:4444::6666:7777:8888',
+        '1111:2222:3333:4444::7777:8888',
+        '1111:2222:3333:4444::8888',
+        '1111:2222:3333::',
+        '1111:2222:3333::5555:6666:123.123.123.123',
+        '1111:2222:3333::5555:6666:7777:8888',
+        '1111:2222:3333::6666:123.123.123.123',
+        '1111:2222:3333::6666:7777:8888',
+        '1111:2222:3333::7777:8888',
+        '1111:2222:3333::8888',
+        '1111:2222::',
+        '1111:2222::4444:5555:6666:123.123.123.123',
+        '1111:2222::4444:5555:6666:7777:8888',
+        '1111:2222::5555:6666:123.123.123.123',
+        '1111:2222::5555:6666:7777:8888',
+        '1111:2222::6666:123.123.123.123',
+        '1111:2222::6666:7777:8888',
+        '1111:2222::7777:8888',
+        '1111:2222::8888',
+        '1111::',
+        '1111::3333:4444:5555:6666:123.123.123.123',
+        '1111::3333:4444:5555:6666:7777:8888',
+        '1111::4444:5555:6666:123.123.123.123',
+        '1111::4444:5555:6666:7777:8888',
+        '1111::5555:6666:123.123.123.123',
+        '1111::5555:6666:7777:8888',
+        '1111::6666:123.123.123.123',
+        '1111::6666:7777:8888',
+        '1111::7777:8888',
+        '1111::8888',
+        '1:2:3:4:5:6:1.2.3.4',
+        '1:2:3:4:5:6:7:8',
+        '1:2:3:4:5:6::',
+        '1:2:3:4:5:6::8',
+        '1:2:3:4:5::',
+        '1:2:3:4:5::7:8',
+        '1:2:3:4:5::8',
+        '1:2:3:4::',
+        '1:2:3:4::5:1.2.3.4',
+        '1:2:3:4::7:8',
+        '1:2:3:4::8',
+        '1:2:3::',
+        '1:2:3::5:1.2.3.4',
+        '1:2:3::7:8',
+        '1:2:3::8',
+        '1:2::',
+        '1:2::5:1.2.3.4',
+        '1:2::7:8',
+        '1:2::8',
+        '1::',
+        '1::2:3',
+        '1::2:3:4',
+        '1::2:3:4:5',
+        '1::2:3:4:5:6',
+        '1::2:3:4:5:6:7',
+        '1::5:1.2.3.4',
+        '1::5:11.22.33.44',
+        '1::7:8',
+        '1::8',
+        '2001:0000:1234:0000:0000:C1C0:ABCD:0876',
+        '2001:0000:4136:e378:8000:63bf:3fff:fdd2',
+        '2001:0db8:0000:0000:0000:0000:1428:57ab',
+        '2001:0db8:0000:0000:0000::1428:57ab',
+        '2001:0db8:0:0:0:0:1428:57ab',
+        '2001:0db8:0:0::1428:57ab',
+        '2001:0db8:1234:0000:0000:0000:0000:0000',
+        '2001:0db8:1234::',
+        '2001:0db8:1234:ffff:ffff:ffff:ffff:ffff',
+        '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
+        '2001:0db8::1428:57ab',
+        '2001::CE49:7601:2CAD:DFFF:7C94:FFFE',
+        '2001::CE49:7601:E866:EFFF:62C3:FFFE',
+        '2001:DB8:0:0:8:800:200C:417A',
+        '2001:DB8::8:800:200C:417A',
+        '2001:db8:85a3:0:0:8a2e:370:7334',
+        '2001:db8:85a3::8a2e:370:7334',
+        '2001:db8::',
+        '2001:db8::1428:57ab',
+        '2001:db8:a::123',
+        '2002::',
+        '2608::3:5',
+        '2608:af09:30:0:0:0:0:134',
+        '2608:af09:30::102a:7b91:c239:baff',
+        '2::10',
+        '3ffe:0b00:0000:0000:0001:0000:0000:000a',
+        '7:6:5:4:3:2:1:0',
+        '::',
+        '::0',
+        '::0:0',
+        '::0:0:0',
+        '::0:0:0:0',
+        '::0:0:0:0:0',
+        '::0:0:0:0:0:0',
+        '::0:0:0:0:0:0:0',
+        '::0:a:b:c:d:e:f',
+        '::1',
+        '::123.123.123.123',
+        '::13.1.68.3',
+        '::2222:3333:4444:5555:6666:123.123.123.123',
+        '::2222:3333:4444:5555:6666:7777:8888',
+        '::2:3',
+        '::2:3:4',
+        '::2:3:4:5',
+        '::2:3:4:5:6',
+        '::2:3:4:5:6:7',
+        '::2:3:4:5:6:7:8',
+        '::3333:4444:5555:6666:7777:8888',
+        '::4444:5555:6666:123.123.123.123',
+        '::4444:5555:6666:7777:8888',
+        '::5555:6666:123.123.123.123',
+        '::5555:6666:7777:8888',
+        '::6666:123.123.123.123',
+        '::6666:7777:8888',
+        '::7777:8888',
+        '::8',
+        '::8888',
+        '::FFFF:129.144.52.38',
+        '::ffff:0:0',
+        '::ffff:0c22:384e',
+        '::ffff:12.34.56.78',
+        '::ffff:192.0.2.128',
+        '::ffff:192.168.1.1',
+        '::ffff:192.168.1.26',
+        '::ffff:c000:280',
+        'FF01:0:0:0:0:0:0:101',
+        'FF01::101',
+        'FF02:0000:0000:0000:0000:0000:0000:0001',
+        'a:b:c:d:e:f:0::',
+        'fe80:0000:0000:0000:0204:61ff:fe9d:f156',
+        'fe80:0:0:0:204:61ff:254.157.241.86',
+        'fe80:0:0:0:204:61ff:fe9d:f156',
+        'fe80::',
+        'fe80::1',
+        'fe80::204:61ff:254.157.241.86',
+        'fe80::204:61ff:fe9d:f156',
+        'fe80::217:f2ff:254.7.237.98',
+        'fe80::217:f2ff:fe07:ed62',
+        'fedc:ba98:7654:3210:fedc:ba98:7654:3210',
+        'ff02::1',
+        'ffff::',
+        'ffff::3:5',
+        'ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff',
+        'a:0::0:b',
+        'a:0:0::0:b',
+        'a:0::0:0:b',
+        'a::0:0:b',
+        'a::0:b',
+        'a:0::b',
+        'a:0:0::b',
+      ]
+      validIpv6.map((ip) => {
+        const valid = isValidIP(ip)
+        assert.equal(valid, true)
+      })
+    })
+  })
+
